// playgo-mobile/src/app/core/api/generated/model/campaignInfo.ts
/**
 * Play&Go Project
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 2.0
 * Contact: info@smartcommunitylab.it
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface CampaignInfo {
  campaignId?: string;
  dateFrom?: number;
  dateTo?: number;
  description?: { [key: string]: string };
  name?: { [key: string]: string };
  startDayOfWeek?: number;
  territoryId?: string;
  type?: CampaignInfo.TypeEnum;
}
export namespace CampaignInfo {
  export type TypeEnum = 'city' | 'company' | 'personal' | 'school';
  export const TypeEnum = {
    City: 'city' as TypeEnum,
    Company: 'company' as TypeEnum,
    Personal: 'personal' as TypeEnum,
    School: 'school' as TypeEnum,
  };
}